Move argument definitions to their own file
